# packages/landing/public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "EmitHQ API",
    "version": "1.0.0",
    "description": "Webhook infrastructure API for sending, receiving, and managing webhooks. Full API-only onboarding — create an account, get an API key, and start sending webhooks without a browser.",
    "contact": { "email": "support@emithq.com", "url": "https://emithq.com" },
    "license": {
      "name": "AGPL-3.0",
      "url": "https://github.com/Not-Another-Ai-Co/EmitHQ/blob/master/LICENSE"
    },
    "x-maintenance": "Hand-written spec. CI checks for route drift via scripts/check-openapi-drift.sh. Update this file when adding or changing API endpoints."
  },
  "servers": [
    { "url": "https://api.emithq.com", "description": "Production" },
    { "url": "http://localhost:4000", "description": "Local development" }
  ],
  "security": [{ "BearerAuth": [] }],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "API key (emhq_ prefix) or Clerk JWT. Pass as: Authorization: Bearer emhq_your_key"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": {
            "type": "object",
            "required": ["code", "message"],
            "properties": {
              "code": { "type": "string", "example": "validation_error" },
              "message": { "type": "string", "example": "name is required" }
            }
          }
        }
      },
...
      "Application": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "uid": { "type": "string", "nullable": true },
          "name": { "type": "string" },
          "createdAt": { "type": "string", "format": "date-time" }
        }
      },
...
    },
    "parameters": {
      "appId": {
        "name": "appId",
        "in": "path",
        "required": true,
        "schema": { "type": "string" },
        "description": "Application UUID or user-defined uid"
      },
...
    },
...
  },
  "paths": {
...
    "/api/v1/app": {
      "post": {
        "operationId": "createApplication",
        "summary": "Create application",
        "tags": ["Applications"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name"],
                "properties": {
                  "name": { "type": "string" },
                  "uid": { "type": "string", "description": "Optional user-defined ID" }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Application created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": { "data": { "$ref": "#/components/schemas/Application" } }
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
            }
          }
        }
      },
      "get": {
        "operationId": "listApplications",
        "summary": "List applications",
        "tags": ["Applications"],
        "responses": {
          "200": {
            "description": "Application list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Application" }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/app/{appId}": {
      "get": {
        "operationId": "getApplication",
        "summary": "Get application",
        "tags": ["Applications"],
        "parameters": [{ "$ref": "#/components/parameters/appId" }],
        "responses": {
          "200": {
            "description": "Application details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": { "data": { "$ref": "#/components/schemas/Application" } }
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
            }
          }
        }
      }
    },
    "/api/v1/app/{appId}/restore": {
      "post": {
        "operationId": "restoreApplication",
        "summary": "Restore a soft-deleted application",
        "description": "Clears the deleted_at timestamp and re-enables endpoints that were disabled by the delete.",
        "tags": ["Applications"],
        "parameters": [{ "$ref": "#/components/parameters/appId" }],
        "responses": {
          "200": {
            "description": "Application restored",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": { "data": { "$ref": "#/components/schemas/Application" } }
                }
              }
            }
          },
          "404": {
            "description": "Deleted application not found",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
            }
          }
        }
      }
    },
...
  },
...
}